refactor: moved dump to cli
