+++ blueprints/tracing/documents/__init__.py
@@ -1,1 +1,2 @@
 from .location import Location
+from .customer import Customer
